# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FrameFinder local inference",
    "description": "FrameFinder 本地向量检索服务；/api/v1 为 ImageLib 正式契约。",
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/capabilities": {
      "get": {
        "summary": "查询 FrameFinder 服务能力",
        "operationId": "capabilities_api_v1_capabilities_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CapabilitiesResponse"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "description": "服务端请求追踪 ID；反馈问题时请提供该值。",
                "schema": {
                  "type": "string",
                  "example": "req-..."
                }
              }
            }
          },
          "400": {
            "description": "请求格式错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "description": "服务端请求追踪 ID；反馈问题时请提供该值。",
                "schema": {
                  "type": "string",
                  "example": "req-..."
                }
              }
            }
          },
          "401": {
            "description": "API Key 无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "description": "服务端请求追踪 ID；反馈问题时请提供该值。",
                "schema": {
                  "type": "string",
                  "example": "req-..."
                }
              }
            }
          },
          "403": {
            "description": "拒绝访问",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "description": "服务端请求追踪 ID；反馈问题时请提供该值。",
                "schema": {
                  "type": "string",
                  "example": "req-..."
                }
              }
            }
          },
          "404": {
            "description": "资产或路径不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "description": "服务端请求追踪 ID；反馈问题时请提供该值。",
                "schema": {
                  "type": "string",
                  "example": "req-..."
                }
              }
            }
          },
          "409": {
            "description": "资产版本或游标状态冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "description": "服务端请求追踪 ID；反馈问题时请提供该值。",
                "schema": {
                  "type": "string",
                  "example": "req-..."
                }
              }
            }
          },
          "413": {
            "description": "上传或批次超过限制",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "description": "服务端请求追踪 ID；反馈问题时请提供该值。",
                "schema": {
                  "type": "string",
                  "example": "req-..."
                }
              }
            }
          },
          "422": {
            "description": "字段或媒体类型不受支持",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "description": "服务端请求追踪 ID；反馈问题时请提供该值。",
                "schema": {
                  "type": "string",
                  "example": "req-..."
                }
              }
            }
          },
          "429": {
            "description": "推理并发已满",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "description": "服务端请求追踪 ID；反馈问题时请提供该值。",
                "schema": {
                  "type": "string",
                  "example": "req-..."
                }
              },
              "Retry-After": {
                "description": "客户端再次尝试前应等待的秒数。",
                "schema": {
                  "type": "integer",
                  "minimum": 0,
                  "example": 1
                }
              }
            }
          },
          "500": {
            "description": "未预期服务端错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "description": "服务端请求追踪 ID；反馈问题时请提供该值。",
                "schema": {
                  "type": "string",
                  "example": "req-..."
                }
              }
            }
          },
          "503": {
            "description": "模型或向量数据库暂不可用",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "description": "服务端请求追踪 ID；反馈问题时请提供该值。",
                "schema": {
                  "type": "string",
                  "example": "req-..."
                }
              }
            }
          },
          "504": {
            "description": "推理超时",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "description": "服务端请求追踪 ID；反馈问题时请提供该值。",
                "schema": {
                  "type": "string",
                  "example": "req-..."
                }
              }
            }
          }
        },
        "security": [
          {
            "FrameFinderApiKey": []
          }
        ]
      }
    },
    "/api/v1/assets/upsert": {
      "post": {
        "summary": "新增或更新单个资产",
        "operationId": "upsert_asset_api_v1_assets_upsert_post",
        "security": [
          {
            "FrameFinderApiKey": []
          }
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upsert_asset_api_v1_assets_upsert_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpsertResponse"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "description": "服务端请求追踪 ID；反馈问题时请提供该值。",
                "schema": {
                  "type": "string",
                  "example": "req-..."
                }
              }
            }
          },
          "400": {
            "description": "请求格式错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "description": "服务端请求追踪 ID；反馈问题时请提供该值。",
                "schema": {
                  "type": "string",
                  "example": "req-..."
                }
              }
            }
          },
          "401": {
            "description": "API Key 无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "description": "服务端请求追踪 ID；反馈问题时请提供该值。",
                "schema": {
                  "type": "string",
                  "example": "req-..."
                }
              }
            }
          },
          "403": {
            "description": "拒绝访问",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "description": "服务端请求追踪 ID；反馈问题时请提供该值。",
                "schema": {
                  "type": "string",
                  "example": "req-..."
                }
              }
            }
          },
          "404": {
            "description": "资产或路径不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "description": "服务端请求追踪 ID；反馈问题时请提供该值。",
                "schema": {
                  "type": "string",
                  "example": "req-..."
                }
              }
            }
          },
          "409": {
            "description": "资产版本或游标状态冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "description": "服务端请求追踪 ID；反馈问题时请提供该值。",
                "schema": {
                  "type": "string",
                  "example": "req-..."
                }
              }
            }
          },
          "413": {
            "description": "上传或批次超过限制",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "description": "服务端请求追踪 ID；反馈问题时请提供该值。",
                "schema": {
                  "type": "string",
                  "example": "req-..."
                }
              }
            }
          },
          "422": {
            "description": "字段或媒体类型不受支持",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "description": "服务端请求追踪 ID；反馈问题时请提供该值。",
                "schema": {
                  "type": "string",
                  "example": "req-..."
                }
              }
            }
          },
          "429": {
            "description": "推理并发已满",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "description": "服务端请求追踪 ID；反馈问题时请提供该值。",
                "schema": {
                  "type": "string",
                  "example": "req-..."
                }
              },
              "Retry-After": {
                "description": "客户端再次尝试前应等待的秒数。",
                "schema": {
                  "type": "integer",
                  "minimum": 0,
                  "example": 1
                }
              }
            }
          },
          "500": {
            "description": "未预期服务端错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "description": "服务端请求追踪 ID；反馈问题时请提供该值。",
                "schema": {
                  "type": "string",
                  "example": "req-..."
                }
              }
            }
          },
          "503": {
            "description": "模型或向量数据库暂不可用",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "description": "服务端请求追踪 ID；反馈问题时请提供该值。",
                "schema": {
                  "type": "string",
                  "example": "req-..."
                }
              }
            }
          },
          "504": {
            "description": "推理超时",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "description": "服务端请求追踪 ID；反馈问题时请提供该值。",
                "schema": {
                  "type": "string",
                  "example": "req-..."
                }
              }
            }
          },
          "200": {
            "description": "资产已更新、复用或仅更新元数据",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpsertResponse"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "description": "服务端请求追踪 ID；反馈问题时请提供该值。",
                "schema": {
                  "type": "string",
                  "example": "req-..."
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upsert_asset_api_v1_assets_upsert_post": {
        "properties": {
          "library_id": {
            "type": "string",
            "maxLength": 256,
            "minLength": 1,
            "title": "Library Id"
          },
          "asset_id": {
            "type": "string",
            "maxLength": 512,
            "minLength": 1,
            "title": "Asset Id"
          },
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          },
          "source_content_hash": {
            "type": "string",
            "title": "Source Content Hash"
          },
          "client_revision": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Client Revision"
          },
          "media_type": {
            "type": "string",
            "enum": [
              "image",
              "video_frame"
            ],
            "title": "Media Type"
          },
          "metadata": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata",
            "description": "JSON 编码的 AssetMetadata 对象；正式字段见 components.schemas.AssetMetadata。"
          },
          "force": {
            "type": "boolean",
            "title": "Force",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "library_id",
          "asset_id",
          "file",
          "source_content_hash",
          "client_revision",
          "media_type"
        ],
        "title": "Body_upsert_asset_api_v1_assets_upsert_post"
      },
      "CapabilitiesResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "api_version": {
            "type": "string",
            "title": "Api Version"
          },
          "server_instance_id": {
            "type": "string",
            "title": "Server Instance Id"
          },
          "accepting_requests": {
            "type": "boolean",
            "title": "Accepting Requests"
          },
          "models_loaded": {
            "type": "boolean",
            "title": "Models Loaded"
          },
          "device": {
            "type": "string",
            "title": "Device"
          },
          "models": {
            "$ref": "#/components/schemas/CapabilityModels"
          },
          "collection": {
            "$ref": "#/components/schemas/CapabilityCollection"
          },
          "supported_media_types": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Supported Media Types"
          },
          "accepted_upload_mime_types": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Accepted Upload Mime Types"
          },
          "supported_filters": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Supported Filters"
          },
          "limits": {
            "$ref": "#/components/schemas/CapabilityLimits"
          },
          "features": {
            "$ref": "#/components/schemas/CapabilityFeatures"
          },
          "server_time": {
            "type": "string",
            "title": "Server Time"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "ok",
          "api_version",
          "server_instance_id",
          "accepting_requests",
          "models_loaded",
          "device",
          "models",
          "collection",
          "supported_media_types",
          "accepted_upload_mime_types",
          "supported_filters",
          "limits",
          "features",
          "server_time",
          "request_id"
        ],
        "title": "CapabilitiesResponse"
      },
      "CapabilityCollection": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "generation": {
            "type": "string",
            "title": "Generation"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "unavailable"
            ],
            "title": "Status"
          },
          "stored_assets": {
            "type": "integer",
            "title": "Stored Assets"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "generation",
          "status",
          "stored_assets"
        ],
        "title": "CapabilityCollection"
      },
...
      "CapabilityModels": {
        "properties": {
          "semantic": {
            "type": "string",
            "title": "Semantic"
          },
          "visual": {
            "type": "string",
            "title": "Visual"
          },
          "model_version": {
            "type": "string",
            "title": "Model Version"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "semantic",
          "visual",
          "model_version"
        ],
        "title": "CapabilityModels"
      },
...
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope"
      },
...
    },
    "securitySchemes": {
      "FrameFinderApiKey": {
        "type": "apiKey",
        "description": "FrameFinder /api/v1 API Key。必须通过安全配置提供，不得硬编码或记录到日志。",
        "in": "header",
        "name": "X-FrameFinder-Key"
      }
    }
  }
}